+++ NEWS.md
@@ -1,5 +1,7 @@
 # spant 2.4.0
 * Performance improvement for reading VD format TWIX files.
+* Function added to scale all metabolite levels by a set value
+(scale_amp_ratio_value).
 
 # spant 2.3.0
 * Verbose option added to read_ima_coil_dir and read_ima_dyn_dir functions.